ecgbackend/requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
sounddevice 
numpy 
torch
openai-whisper
torchaudio
openai-whisper
python-multipart
gradio